+++ src/lib.rs
@@ -73,9 +73,9 @@ pub use crate::element::Element;
 pub use crate::error::Result;
 #[cfg(feature = "fetcher")]
 pub use crate::fetcher::{BrowserFetcher, BrowserFetcherOptions};
-pub use crate::handler::{ClickOptions, ClickOptionsBuilder};
 pub use crate::handler::Handler;
 pub use crate::handler::movement::MovementBehavior;
+pub use crate::handler::{ClickOptions, ClickOptionsBuilder};
 pub use crate::page::Page;
 
 pub mod types {