+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `scroll` method to `Page`
 - Add simulated mouse movement
 - Add `human_movements` feature for human like mouse movement and click jitter
 - Add `mouse_pos` method to `Page`